+++ tslib/react/.storybook/main.ts
@@ -2,18 +2,19 @@ import type { StorybookConfig } from "@storybook/react-vite"
 
 const config: StorybookConfig = {
   stories: ["../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
+
   addons: [
     "@storybook/addon-links",
     "@storybook/addon-essentials",
     "@storybook/addon-interactions",
+    "@chromatic-com/storybook",
   ],
+
   framework: {
     name: "@storybook/react-vite",
     options: {},
   },
-  docs: {
-    autodocs: "tag",
-  },
+
   async viteFinal(config) {
     const { mergeConfig } = await import("vite")
     return mergeConfig(config, {
@@ -24,6 +25,10 @@ const config: StorybookConfig = {
         },
       },
     })
+  },
+
+  typescript: {
+    reactDocgen: "react-docgen-typescript",
   },
 }
 export default config
