+++ tslib/react/package.json
@@ -38,15 +38,16 @@
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@optuna/types": "file:../types",
+    "@chromatic-com/storybook": "^1.6.1",
     "@optuna/storage": "file:../storage",
-    "@storybook/addon-essentials": "^8.0.4",
-    "@storybook/addon-interactions": "^8.0.4",
-    "@storybook/addon-links": "^8.0.4",
-    "@storybook/blocks": "^8.0.4",
-    "@storybook/react": "^8.0.4",
-    "@storybook/react-vite": "^8.0.4",
-    "@storybook/test": "^8.0.4",
+    "@optuna/types": "file:../types",
+    "@storybook/addon-essentials": "^8.2.9",
+    "@storybook/addon-interactions": "^8.2.9",
+    "@storybook/addon-links": "^8.2.9",
+    "@storybook/blocks": "^8.2.9",
+    "@storybook/react": "^8.2.9",
+    "@storybook/react-vite": "^8.2.9",
+    "@storybook/test": "^8.2.9",
     "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/react": "^14.2.2",
     "@types/plotly.js-dist-min": "^2.3.4",
@@ -54,7 +55,7 @@
     "@types/react-dom": "^18.2.19",
     "@vitejs/plugin-react-swc": "^3.5.0",
     "jsdom": "^24.0.0",
-    "storybook": "^8.0.4",
+    "storybook": "^8.2.9",
     "typescript": "^5.2.2",
     "vite": "^5.1.0",
     "vitest": "^1.4.0"
