+++ tslib/react/tsconfig.json
@@ -23,12 +23,11 @@
     "noFallthroughCasesInSwitch": true,
     "jsx": "react-jsx"
   },
-  "include": ["src/**/*.ts", "src/**/*.tsx"],
+  "include": ["src/**/*.ts", "src/**/*.tsx", ".storybook/**/*.tsx"],
   "exclude": [
     "src/**/*.stories.tsx",
     "src/**/*.stories.ts",
     "src/MockStudies.ts",
-    "src/utils/loadStorageFromFile.ts"
   ],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
